mykernel2.c

+++ pa2/mykernel2.c
@@ -32,7 +32,7 @@ void set_requested_ratio(int pid, int m, int n){
       return;
     }
     else if(proctab[i].pid == pid) {
-      proctab[i].requested = m / n;
+      proctab[i].requested = (double)m / n;
       Printf("set %d requested %f \n", pid, m/n);
       return;
     }
